refactor: extract llm_base and analysis_models from agent_responses
agent_responses.py had grown to hold three separate concerns. Split it into
focused modules:

- agents/llm_base.py        — the LLMBaseModel schema/field-hiding framework
- agents/agent_responses.py — only the LLM-facing response schemas
- agents/analysis_models.py — the runtime/domain models (Component,
  AnalysisInsights, Relation, MethodEntry, FileMethodGroup, FileEntry) and the
  component-tree helpers (assign_component_ids, iter_components,
  index_components_by_id)

Also drops the dead CFGComponent / CFGAnalysisInsights. Imports across the
codebase are repointed to the new modules; no behavior change.
