+++ agents/details_agent.py
@@ -5,14 +5,8 @@ from langchain_core.prompts import PromptTemplate
 from langchain_core.language_models import BaseChatModel
 
 from agents.agent import CodeBoardingAgent
-from agents.agent_responses import (
-    AnalysisInsights,
-    AnalysisInsightsLLM,
-    ClusterAnalysis,
-    Component,
-    MetaAnalysisInsights,
-    assign_component_ids,
-)
+from agents.agent_responses import AnalysisInsightsLLM, ClusterAnalysis, MetaAnalysisInsights
+from agents.analysis_models import AnalysisInsights, Component, assign_component_ids
 from agents.prompts import get_system_details_message, get_cfg_details_message, get_details_message
 from agents.cluster_methods_mixin import ClusterMethodsMixin
 from caching.cache import ModelSettings
