+++ diagram_analysis/analysis_json.py
@@ -4,15 +4,8 @@ from datetime import datetime, timezone
 
 from pydantic import BaseModel, Field
 
-from agents.agent_responses import (
-    Component,
-    Relation,
-    AnalysisInsights,
-    FileEntry,
-    FileMethodGroup,
-    MethodEntry,
-    SourceCodeReference,
-)
+from agents.agent_responses import SourceCodeReference
+from agents.analysis_models import Component, Relation, AnalysisInsights, FileEntry, FileMethodGroup, MethodEntry
 
 logger = logging.getLogger(__name__)
