+++ output_generators/html.py
@@ -2,7 +2,7 @@ from pathlib import Path
 from typing import Dict, Any
 import json
 
-from agents.agent_responses import AnalysisInsights
+from agents.analysis_models import AnalysisInsights
 from utils import sanitize
 from output_generators.html_template import populate_html_template
