+++ tests/output_generators/test_output_generators.py
@@ -3,13 +3,8 @@ import unittest
 from pathlib import Path
 from unittest.mock import patch
 
-from agents.agent_responses import (
-    AnalysisInsights,
-    Component,
-    Relation,
-    SourceCodeReference,
-    assign_component_ids,
-)
+from agents.agent_responses import SourceCodeReference
+from agents.analysis_models import AnalysisInsights, Component, Relation, assign_component_ids
 from utils import sanitize
 from output_generators.html import (
     component_header_html,
